Move vote counting to database query

+++ src/main/java/com/github/denconcept/restaurantvoting/restaurant/to/AdminResponseVotingResultTo.java
@@ -3,5 +3,5 @@ package com.github.denconcept.restaurantvoting.restaurant.to;
 public record AdminResponseVotingResultTo(
         Integer restaurantId,
         String name,
-        Integer voteCount) {
+        Long voteCount) {
 }
